feat(依赖): 升级所有依赖到最新版本

+++ internal/config/snowflake.go
@@ -2,5 +2,5 @@ package config
 
 type Snowflake struct {
 	// 节点
-	Node int `default:"1" json:"node" yaml:"node" xml:"node" toml:"node" validate:"required,min=1,max=1024"`
+	Node uint16 `default:"1" json:"node" yaml:"node" xml:"node" toml:"node" validate:"required,min=1,max=1024"`
 }
